+++ src/appshell/qml/Preferences/internal/MeterDbRangeSection.qml
@@ -13,8 +13,12 @@ BaseSection {
     title: qsTrc("appshell/preferences", "Meter dB range")
     spacing: 16
 
-    PlaybackMeterModel {
-        id: playbackMeterModel
+    MeterModel {
+        id: meterModel
+    }
+
+    Component.onCompleted: {
+        meterModel.init();
     }
 
     ComboBoxWithTitle {
@@ -22,9 +26,9 @@ BaseSection {
 
         columnWidth: 320
 
-        currentIndex: playbackMeterModel.meterDbRange
-        model: playbackMeterModel.dbRangeList.map(function(id) {
-            return playbackMeterModel.description(id);
+        currentIndex: meterModel.meterDbRange
+        model: meterModel.dbRangeList.map(function(id) {
+            return meterModel.description(id);
         })
 
         navigation.name: "MeterDbRangeBox"
@@ -32,7 +36,7 @@ BaseSection {
         navigation.row: 1
 
         onValueEdited: function(newIndex, _) {
-            playbackMeterModel.meterDbRange = newIndex;
+            meterModel.meterDbRange = newIndex;
         }
     }
 }
