+++ src/record/recordmodule.cpp
@@ -14,6 +14,7 @@
 #include "internal/recordcontroller.h"
 #include "internal/recorduiactions.h"
 #include "internal/au3/au3record.h"
+#include "view/common/recordmetermodel.h"
 
 using namespace au::record;
 using namespace muse;
@@ -54,6 +55,11 @@ void RecordModule::resolveImports()
 void RecordModule::registerResources()
 {
     record_init_qrc();
+}
+
+void RecordModule::registerUiTypes()
+{
+    qmlRegisterType<RecordMeterModel>("Audacity.Record", 1, 0, "RecordMeterModel");
 }
 
 void RecordModule::onInit(const IApplication::RunMode& mode)
