+++ src/record/recordmodule.h
@@ -21,6 +21,7 @@ public:
     void registerExports() override;
     void resolveImports() override;
     void registerResources() override;
+    void registerUiTypes() override;
     void onInit(const muse::IApplication::RunMode& mode) override;
     void onDeinit() override;
